refactor: seperate add file into three directories
