+++ sprint7/package.json
@@ -3,8 +3,6 @@
     "@prisma/client": "^5.4.2",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "is-email": "^1.0.2",
-    "is-uuid": "^1.0.2",
     "prisma": "^5.4.2",
     "superstruct": "^1.0.3"
   },
